refactor(transcription): address second-round diff-review findings
- Compute effectiveDiarize once per run() and reuse it for both the
  provider's diarize option and the speaker-stripping decision, so the
  request-time and output-time gates are visibly the same value and cannot
  diverge.
- Extract the capturing Obsidian Setting test double into a shared
  tests/helpers/captureSettings module, replacing the duplicated inline
  mocks in settingControls.test.ts and transcriptionSettingsSection.test.ts.
  The global tests/mocks/obsidian.ts stays unusable here because its builder
  methods never invoke the callback, so a capturing variant is still needed
  — now in one place.
