+++ src/transcription/TranscriptionService.ts
@@ -123,20 +123,22 @@ export class TranscriptionService {
 		const settings = this.getSettings();
 		const token = options.token ?? NEVER_CANCELLED;
 		const provider = this.createProvider(settings);
+		// One gate for the whole run: a stale "on" left from a diarizing engine
+		// is ignored and a non-diarizing engine never sends a field it would
+		// silently drop. The same value decides both whether to request speaker
+		// labels here and whether to strip any the provider returned (below), so
+		// the request-time and output-time decisions can never diverge.
+		const diarize = effectiveDiarize(
+			settings.transcriptionProvider,
+			settings.transcriptionDiarize,
+		);
 		const transcribeOptions = {
 			language:
 				settings.transcriptionLanguage &&
 				settings.transcriptionLanguage !== 'auto'
 					? settings.transcriptionLanguage
 					: undefined,
-			// Only request diarization when the engine can actually produce
-			// speaker labels. effectiveDiarize is the shared gate, so a stale
-			// "on" left from a diarizing engine is ignored and a non-diarizing
-			// engine never sends a field it would silently drop.
-			diarize: effectiveDiarize(
-				settings.transcriptionProvider,
-				settings.transcriptionDiarize,
-			),
+			diarize,
 			wordTimestamps: settings.transcriptionWordTimestamps,
 		};
 
@@ -230,12 +232,7 @@ export class TranscriptionService {
 		// JSON) shows a label the user did not ask for. Doing it once here, on
 		// the canonical transcript, keeps every consumer consistent rather than
 		// gating each renderer separately.
-		const transcript = effectiveDiarize(
-			settings.transcriptionProvider,
-			settings.transcriptionDiarize,
-		)
-			? stitched
-			: stripSpeakers(stitched);
+		const transcript = diarize ? stitched : stripSpeakers(stitched);
 
 		const markdownOptions = this.markdownOptions(settings);
 		let markdown = formatTranscriptMarkdown(
