+++ tests/unit/settingControls.test.ts
@@ -2,8 +2,9 @@
  * Unit tests for the shared setting-control builders, focused on the
  * disabled (dimmed) rendering used for options the current selection cannot
  * use — e.g. speaker diarization on an engine that cannot diarize. A disabled
- * toggle must both be non-interactive and visually greyed (a dim class on the
+ * control must both be non-interactive and visually greyed (a dim class on the
  * row), so the user can tell it is off because it is unavailable, not unset.
+ * The capturing Setting mock is shared from tests/helpers/captureSettings.
  * @module tests/unit/settingControls.test
  */
 /** @jest-environment jsdom */
@@ -15,95 +16,12 @@ import {
 	type SettingsSectionContext,
 } from '../../src/settings/settingControls';
 import type { AudioRecorderSettings } from '../../src/settings/Settings';
-
-/** One rendered setting captured through the Setting mock. */
-interface SettingCapture {
-	name: string;
-	el: HTMLElement;
-	toggle: { value: boolean; disabled: boolean } | null;
-	text: { value: string; disabled: boolean } | null;
-}
-const captured: SettingCapture[] = [];
+import { capturedSettings } from '../helpers/captureSettings';
 
 jest.mock('obsidian', () => ({
-	Setting: class {
-		settingEl: HTMLElement;
-		private cap: SettingCapture;
-		constructor(_containerEl: HTMLElement) {
-			const el = document.createElement('div');
-			(el as unknown as { addClass: (c: string) => void }).addClass = (
-				c,
-			) => el.classList.add(c);
-			this.settingEl = el;
-			this.cap = { name: '', el, toggle: null, text: null };
-			captured.push(this.cap);
-		}
-		setName(name: string): this {
-			this.cap.name = name;
-			return this;
-		}
-		setDesc(): this {
-			return this;
-		}
-		addText(
-			callback: (text: {
-				value: string;
-				disabled: boolean;
-				inputEl: { type: string };
-				setValue: (v: string) => unknown;
-				onChange: (h: (v: string) => void) => unknown;
-				setDisabled: (d: boolean) => unknown;
-			}) => void,
-		): this {
-			const text = {
-				value: '',
-				disabled: false,
-				inputEl: { type: 'text' },
-				setValue(v: string) {
-					this.value = v;
-					return this;
-				},
-				onChange() {
-					return this;
-				},
-				setDisabled(d: boolean) {
-					this.disabled = d;
-					return this;
-				},
-			};
-			callback(text);
-			this.cap.text = text;
-			return this;
-		}
-		addToggle(
-			callback: (toggle: {
-				value: boolean;
-				disabled: boolean;
-				setValue: (v: boolean) => unknown;
-				onChange: (h: (v: boolean) => void) => unknown;
-				setDisabled: (d: boolean) => unknown;
-			}) => void,
-		): this {
-			const toggle = {
-				value: false,
-				disabled: false,
-				setValue(v: boolean) {
-					this.value = v;
-					return this;
-				},
-				onChange() {
-					return this;
-				},
-				setDisabled(d: boolean) {
-					this.disabled = d;
-					return this;
-				},
-			};
-			callback(toggle);
-			this.cap.toggle = toggle;
-			return this;
-		}
-	},
+	Setting: jest.requireActual<typeof import('../helpers/captureSettings')>(
+		'../helpers/captureSettings',
+	).CapturingSetting,
 }));
 
 /** Builds a section context whose hooks are spies. */
@@ -119,7 +37,7 @@ function makeCtx(): SettingsSectionContext {
 
 describe('addToggle disabled rendering', () => {
 	beforeEach(() => {
-		captured.length = 0;
+		capturedSettings.length = 0;
 	});
 
 	it('dims the row and disables the toggle when disabled', () => {
@@ -130,7 +48,7 @@ describe('addToggle disabled rendering', () => {
 			disabled: true,
 		});
 
-		const setting = captured[0];
+		const setting = capturedSettings[0];
 		expect(setting.el.classList.contains(SETTING_DISABLED_CLASS)).toBe(
 			true,
 		);
@@ -144,7 +62,7 @@ describe('addToggle disabled rendering', () => {
 			set: () => undefined,
 		});
 
-		const setting = captured[0];
+		const setting = capturedSettings[0];
 		expect(setting.el.classList.contains(SETTING_DISABLED_CLASS)).toBe(
 			false,
 		);
@@ -155,7 +73,7 @@ describe('addToggle disabled rendering', () => {
 
 describe('addText disabled rendering', () => {
 	beforeEach(() => {
-		captured.length = 0;
+		capturedSettings.length = 0;
 	});
 
 	it('dims the row and disables the input when disabled', () => {
@@ -166,7 +84,7 @@ describe('addText disabled rendering', () => {
 			disabled: true,
 		});
 
-		const setting = captured[0];
+		const setting = capturedSettings[0];
 		expect(setting.el.classList.contains(SETTING_DISABLED_CLASS)).toBe(
 			true,
 		);
@@ -180,7 +98,7 @@ describe('addText disabled rendering', () => {
 			set: () => undefined,
 		});
 
-		const setting = captured[0];
+		const setting = capturedSettings[0];
 		expect(setting.el.classList.contains(SETTING_DISABLED_CLASS)).toBe(
 			false,
 		);
